Move CSR module
