+++ wake_t/__init__.py
@@ -1,7 +1,7 @@
 from .__version__ import __version__
 from .beamline_elements import (PlasmaStage, PlasmaRamp, PlasmaLens, Drift,
                                 Dipole, Quadrupole, Sextupole, Beamline)
-from .csr import set_csr_settings
+from .physics_models.collective_effects.csr import set_csr_settings
 
 __all__ = ['__version__', 'PlasmaStage', 'PlasmaRamp', 'PlasmaLens', 'Drift',
            'Dipole', 'Quadrupole', 'Sextupole', 'Beamline', 'set_csr_settings']
